requirements.txt
pytest==8.0.2
pytest-pycodestyle
requests
flask
Pillow
